Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
regex==2023.10.3
PyPDF2==3.0.1
reportlab==4.0.4
python-magic==0.4.27
aiofiles==23.2.1
pdfplumber==0.10.3
pymupdf==1.23.8
pytesseract==0.3.13
pillow==11.2.1 